Use SingleApplication component for both public sharing and private view/edit
